Update Yob.rb
Some adjustments where made in the code, in order to make the program even more user-friendly. This program is currently in Beta-Status, and needs adjustments in the part of the code that data process the year of birth (please kindly check line 50) in order to be fully functional. Such adjustments will be troubleshooted asap.